done with lets get hooked
