+++ index.css
@@ -44,3 +44,9 @@
     display: flex;
     flex-wrap: wrap;
 }
+
+.filter-btn{
+    cursor: pointer;
+    margin: 10px;
+}
+
