+++ index.html
@@ -11,7 +11,7 @@
         <h1>Not Rendered!</h1>
     </div>
 
-    <script type="module" src="./App.js">
+    <script type="module" src="./src/App.js">
         
     </script>
