+++ src/components/Body.js
@@ -1,0 +1,32 @@
+import RestaurantCard from "./RestaurantCard";
+import { useState } from "react";
+import resList from "../utils/mockData";
+
+const Body = () => {
+  const [listofRestaurants , setlistofRestaurant] = useState(resList);
+  return (
+    <div className="body">
+      <div className="filter">
+        <button
+          className="filter-btn"
+          onClick={() => {
+            // Filter listofRestaurants
+            let filteredRestaurants = listofRestaurants.filter(
+              (res) => res.info.avgRating > 4
+            );
+            setlistofRestaurant(filteredRestaurants);
+          }}
+        >
+          Top Rated Restaurants
+        </button>
+      </div>
+      <div className="res-container">
+        {listofRestaurants.map((restaurant) => (
+          <RestaurantCard key={restaurant.info.id} resData={restaurant} />
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default Body;
