+++ src/utils/constants.js
@@ -1,0 +1,5 @@
+export const CDN_URL = 
+"https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
+export const LOGO_URL =
+"https://img.freepik.com/free-vector/restaurant-tasty-food-logo-design_460848-10307.jpg?w=360&t=st=1697882039~exp=1697882639~hmac=fcf364dbf9ea0f8cf717dbdd371f46accb6669ace7cb969e4388324794e2ee0e";
+
